requirements.txt
asgiref==3.6.0
Django==3.2
django-cleanup==7.0.0
django-filter==23.1
django-grappelli==3.0.6
django-tinymce==3.6.1
djangorestframework==3.14.0
isort==5.12.0
Markdown==3.4.3
Pillow==9.5.0
pytz==2023.3
sorl-thumbnail==12.9.0
sqlparse==0.4.3
